Update report_routes.js
deleted queries in the list for the various reports in cluded in this js so that it matches those that are present/available within the interface....fixed duplicate query ids in the list and those that aren't yet implemented.  Wondering about ordering and if the list of query ids also imparts list order within the report - have a question out to Simon for this.